Enable full 64-bit Linux support for xrCore and Filesystem
- Updated IReader, IWriter, and CLocatorAPI to use size_t for offsets and sizes, enabling support for files > 4GB.
- Replaced 32-bit pointer casts with uintptr_t in memory management and filesystem logic to prevent truncation.
- Fixed xrMEMORY_POOL to correctly handle 64-bit pointers during block partitioning.
- Updated memory allocation shims (xr_aligned_msize, pso_MemFill/Copy) to use size_t.
- Implemented malloc_usable_size as a Linux replacement for _msize in xrMemory_align.cpp.
- Fixed GetFileSize and SetFilePointer shims to support 64-bit offsets and prevent overflows.
- Refined SEH shims and _beginthread implementation for better C++ compatibility and stability on Linux.
- Ensured proper HRESULT definitions and signedness for 64-bit architectures.

+++ src/xrCore/LocatorAPI.h
@@ -26,9 +26,9 @@ public:
 		LPCSTR name; // low-case name
 		u32 vfs; // 0xffffffff - standart file
 		u32 crc; // contents CRC
-		u32 ptr; // pointer inside vfs
-		u32 size_real; //
-		u32 size_compressed; // if (size_real==size_compressed) - uncompressed
+		size_t ptr; // pointer inside vfs
+		size_t size_real; //
+		size_t size_compressed; // if (size_real==size_compressed) - uncompressed
 		u32 modif; // for editor
 	};
 
@@ -36,7 +36,7 @@ public:
 	{
 		shared_str path;
 		void *hSrcFile, *hSrcMap;
-		u32 size;
+		size_t size;
 		CInifile* header;
 		u32 vfs_idx;
 
@@ -78,7 +78,7 @@ private:
 	xrCriticalSection m_auth_lock;
 	u64 m_auth_code;
 
-	void Register(LPCSTR name, u32 vfs, u32 crc, u32 ptr, u32 size_real, u32 size_compressed, u32 modif);
+	void Register(LPCSTR name, u32 vfs, u32 crc, size_t ptr, size_t size_real, size_t size_compressed, u32 modif);
 	void ProcessArchive(LPCSTR path);
 	void ProcessOne(LPCSTR path, const _FINDDATA_T& entry);
 	bool Recurse(LPCSTR path);
