+++ src/xrCore/xrMemory.cpp
@@ -274,7 +274,7 @@ void xrMemory::mem_statistic(LPCSTR fn)
 char* xr_strdup(const char* string)
 {
 	VERIFY(string);
-	u32 len = u32(xr_strlen(string)) + 1;
+	size_t len = xr_strlen(string) + 1;
 	char* memory = (char*)Memory.mem_alloc(len
 #ifdef DEBUG_MEMORY_NAME
                                            , "strdup"
@@ -289,6 +289,6 @@ XRCORE_API BOOL is_stack_ptr(void* _ptr)
 	int local_value = 0;
 	void* ptr_refsound = _ptr;
 	void* ptr_local = &local_value;
-	ptrdiff_t difference = (ptrdiff_t)_abs(s64(ptrdiff_t(ptr_local) - ptrdiff_t(ptr_refsound)));
+	ptrdiff_t difference = (ptrdiff_t)_abs(s64((intptr_t)ptr_local - (intptr_t)ptr_refsound));
 	return (difference < (512 * 1024));
 }
